pyproject.toml
[tool.ruff]

# description of all rules are available on https://docs.astral.sh/ruff/rules/
lint.select = ["D", "E", "F", "W", "C", "S", "I", "TCH", "SLOT", "RUF", "C90", "N", "YTT", "ASYNC", "TRIO"]

# we need to check 'mood' of all docstrings, this needs to be enabled explicitly
lint.extend-select = ["D401"]

lint.ignore = []

target-version = "py311"
lint.pydocstyle.convention = "google"
line-length = 100

[tool.mypy]
disable_error_code = ["union-attr", "return-value", "arg-type", "import-untyped"]
ignore_missing_imports = true

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "black==24.4.2",
    "mypy==1.10.0",
    "ruff==0.4.6",
    "types-requests==2.31.0"
]

[[tool.pdm.source]]
type = "find_links"
url = "https://download.pytorch.org/whl/cpu/torch_stable.html"
name = "torch"

[project]
name = "lightspeed-rag-content"
version = "0.1.0"
description = "RAG content for OpenShift LightSpeed."
authors = []
dependencies = [
    "PyYAML==6.0.1",
    "huggingface_hub==0.23.2",
    "llama_index==0.10.43",
    "llama-index-vector-stores-faiss==0.1.2",
    "llama-index-embeddings-huggingface==0.2.1",
    "faiss-cpu==1.8.0",
    "spacy==3.7.4",  
    "nest-asyncio==1.6.0", 
    "langchain==0.1.16", 
    "langchain-community==0.0.34",
    "llama-index-llms-langchain==0.1.3",
    "python-dotenv==1.0.0",
    "langchain-ibm==0.1.3",
    "ibm-watson-machine-learning==1.0.354",
    "ibm-generative-ai==2.3.0",
    "langchain-openai==0.1.1",
    "transformers==4.40.1",
    "sentence-transformers==2.7.0"
]
requires-python = "==3.11.*"
readme = "README.md"
license = "LICENSE"
